Female Breast Cancer Tracking System GP
Project Background:
The choice of female breast cancer tracking system is due to the great
appearance of the
female breast cancer. These days many hospitals and centers around the
world are interested in treating this type of disease one of these
hospitals is Baheya hospital. Baheya hospital is one of charity
hospitals in Egypt, it specialized in the treatment of the breast cancer
for women only, Baheya is established in order to help every woman
detect the cancer early to be free from breast cancer, and to make the
services of the hospital easier and flexible to patients and doctors we
had to update the Baheya‟s website to be more dynamic instead of only
displays its information and events.

Scope:
Baheya„s website will be renewed to include more features that may
concern both patients and doctors, some of these features are:
1- Making the reservation process for patients easier and allow patients
to view their medical history.
2- Doctors will use the website to keep track of patients who are
assigned to them by viewing their medical history, write a new
prescription…etc.
3- The website also allow all system‟s users to communicate with each
other as patients can send messages to their doctors if they need to ask
anything about their examination, treatments and so on.
4- Moreover, if a set of doctors have the same case they can also
communicate with each other, sharing medical cases, leaving comments on
each other‟s cases.

For more information about the system, read the doc "Documentation.pdf"
Screen shots of the developed application can be found pages (29:48) at
the doc Documentation.pdf